+++ src/main/java/baseball/view/ConsoleOutput.java
@@ -1,4 +1,4 @@
-package baseball.io;
+package baseball.view;
 
 import java.io.BufferedWriter;
 import java.io.IOException;
